Issue 9: Add unit tests for games
Adding unit tests infrastructure, and basic unit tests for dictionary
and word search. Also fixing accessibility level of some classes in the
main project for testability.

These are the first tests. I'll add tests as necessary in the future.
Philosophy will be to add tests with new components or add tests that
cover bugs that are found so we don't regress them.